API Functions & Little changes + New Patches
Added patch to another game that lags my PC, and added a few new function to External LuaU API such as WaitFor(objectString) that will return an object with waitforchilds, and Rejoin() that will rejoin the server. I've tried to test around with Church update with it's Adonis commands, and still can't seem to get it, but I do know a way to not show commands which is through the allies. I will later make something to make allies automatically and you can use them.